Develop

See merge request ckb-explorer/ckb-explorer-web!161

+++ src/main/java/com/ckb/explorer/domain/dto/CellOutputDto.java
@@ -1,5 +1,6 @@
 package com.ckb.explorer.domain.dto;
 
+import java.math.BigInteger;
 import java.util.List;
 import lombok.Data;
 
@@ -9,12 +10,12 @@ public class CellOutputDto {
   /**
    * 单元格容量
    */
-  private Long capacity;
+  private BigInteger capacity;
 
   /**
    * 占用的容量
    */
-  private Long occupiedCapacity;
+  private BigInteger occupiedCapacity;
 
   private byte[] lockCodeHash;
